requirements.txt
Flask
Mysql-connector-python
Pandas